Moved Flask api to `app.py` to make api calls work
